requirements.txt
appdirs==1.4.4
asgiref==3.3.4
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.12
distlib==0.3.2
Django
django-filter==2.4.0
django-phone-verify==2.0.1
django-phonenumber-field==5.2.0
django-rest-authemail==2.0.6
djangorestframework
djangorestframework-simplejwt==4.7.1
drf-yasg==1.20.0
filelock==3.0.12
google-api-core==1.30.0
google-api-python-client==2.10.0
google-auth==1.32.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
Markdown==3.3.4
MarkupSafe==2.0.1
nexmo==2.5.2
packaging==20.9
phonenumbers==8.12.25
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.1.0
pyparsing==2.4.7
python-decouple==3.4
python-dotenv==0.18.0
pytz==2021.1
requests==2.25.1
rsa==4.7.2
ruamel.yaml==0.17.9
ruamel.yaml.clib==0.2.2
six==1.16.0
sqlparse==0.4.1
twilio==6.60.0
uritemplate==3.0.1
urllib3==1.26.5
virtualenv==20.4.7
virtualenv-clone==0.5.4
whitenoise==5.2.0
wrapt==1.12.1
